chore: add e2e for SDS Ref (#8983)
* add more sds e2e


* merge test suite


* fix merge


---------
